+++ apps/sample-ehr/src/pages/patients/[patientId]/edit.tsx
@@ -1,8 +1,8 @@
 import { MainPage } from "@/components";
 import { PatientIdentifierSystems } from "@/fhir/known-identifiers";
 import { CustomPatient } from "@/fhir/patient";
-import { useFhirResourceForm } from "@bonfhir/ui-mantine/r4b";
-import { FhirInput, FhirInputArray } from "@bonfhir/ui/r4b";
+import { useFhirResourceForm } from "@bonfhir/mantine/r4b";
+import { FhirInput, FhirInputArray } from "@bonfhir/react/r4b";
 import {
   Box,
   Button,