+++ packages/next/CHANGELOG.md
@@ -31,7 +31,7 @@
 - Updated dependencies [[`6b8e816`](https://github.com/bonfhir/bonfhir/commit/6b8e8164afea6c06de22bf8e1313b29057a9ff6e), [`e5dd1d5`](https://github.com/bonfhir/bonfhir/commit/e5dd1d5411f4ae68ecff706f2f0277ab766e7aac), [`3d1ce25`](https://github.com/bonfhir/bonfhir/commit/3d1ce25cbc26d6b272f1388fd3210abea52ac50e), [`623cac8`](https://github.com/bonfhir/bonfhir/commit/623cac852d3f84ff5209282069a0d1d95a8b30cc)]:
   - @bonfhir/subscriptions@2.2.0
   - @bonfhir/core@2.8.0
-  - @bonfhir/ui@2.3.0
+  - @bonfhir/react@2.3.0
 
 ## 2.1.1
 
@@ -43,7 +43,7 @@
 - Updated dependencies [7891d57]
   - @bonfhir/core@2.6.0
   - @bonfhir/subscriptions@2.1.1
-  - @bonfhir/ui@2.1.1
+  - @bonfhir/react@2.1.1
 
 ## 2.1.0
 
@@ -56,4 +56,4 @@
 - Updated dependencies
   - @bonfhir/core@2.1.0
   - @bonfhir/subscriptions@2.1.0
-  - @bonfhir/ui@2.1.0
+  - @bonfhir/react@2.1.0